pyproject.toml
[project]
name = "pracitice-ichibero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "rio-tiler>=7.6.0",
]
